requirements.txt
Brotli==1.0.9
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.0
lxml==4.6.2
MarkupSafe==1.1.1
multitasking==0.0.9
nltk==3.5
numpy==1.20.0
pandas==1.2.1
pip==20.1.1
pkg-resources==0.0.0
plotly==4.14.3
praw==7.1.2
prawcore==1.5.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
regex==2020.11.13
requests==2.25.1
retrying==1.3.3
setuptools==44.0.0
six==1.15.0
textblob==0.15.3
tqdm==4.56.0
update-checker==0.18.0
urllib3==1.26.3
websocket-client==0.57.0
Werkzeug==1.0.1
yfinance==0.1.55
